Moved validators
