+++ lib/specs.js
@@ -31,7 +31,7 @@ var JsonRefs = require('json-refs');
 var SparkMD5 = require('spark-md5');
 var swaggerConverter = require('swagger-converter');
 var traverse = require('traverse');
-var validators = require('./validators');
+var validators = require('../src/lib/validators');
 var YAML = require('js-yaml');
 
 // Work around swagger-converter packaging issue (Browser builds only)
